requirements.txt
flask
flask-wtf
google-generativeai
